+++ front/src/App/pages/SongRegister/component/renderField.js
@@ -44,3 +44,26 @@ export const renderAreaField = ({
     </div>
   </div>
 )
+
+export const renderCheckBoxField = ({
+  input,
+  label,
+  type,
+  meta: { touched, error, warning }
+}) => (
+  <div>
+    <div style={{ display: "flex" }}>
+      <input
+        {...input}
+        type={type}
+        className="form-control"
+        value={input.value}
+        id={label}
+      />
+      <label htmlFor={label}>{label}</label>
+      {touched &&
+        ((error && <span>{error}</span>) ||
+          (warning && <span>{warning}</span>))}
+    </div>
+  </div>
+)
