+++ front/src/App/pages/SongRegister/component/textInput.js
@@ -5,56 +5,54 @@ class TextInputForm extends PureComponent {
   render() {
     return (
       <div>
-        <form>
-          <div className="form-group">
-            <Field
-              type="text"
-              name="title"
-              label="Title"
-              component={renderInputField}
-            />
-          </div>
-          <div className="form-group">
-            <Field
-              type="text"
-              name="Feat"
-              label="Feat"
-              component={renderInputField}
-            />
-          </div>
-          <div className="form-group">
-            <Field
-              type="text"
-              name="Upload"
-              label="Upload"
-              component={renderInputField}
-            />
-          </div>
-          <div className="form-group">
-            <Field
-              type="text"
-              name="Lyrics"
-              label="Lyrics"
-              component={renderAreaField}
-            />
-          </div>
-          <div className="form-group">
-            <Field
-              type="text"
-              name="Description"
-              label="Description"
-              component={renderAreaField}
-            />
-          </div>
-          <div className="form-group">
-            <Field
-              type="text"
-              name="Email"
-              label="Email"
-              component={renderInputField}
-            />
-          </div>
-        </form>
+        <div className="form-group">
+          <Field
+            type="text"
+            name="title"
+            label="Title"
+            component={renderInputField}
+          />
+        </div>
+        <div className="form-group">
+          <Field
+            type="text"
+            name="Feat"
+            label="Feat"
+            component={renderInputField}
+          />
+        </div>
+        <div className="form-group">
+          <Field
+            type="text"
+            name="Upload"
+            label="Upload"
+            component={renderInputField}
+          />
+        </div>
+        <div className="form-group">
+          <Field
+            type="text"
+            name="Lyrics"
+            label="Lyrics"
+            component={renderAreaField}
+          />
+        </div>
+        <div className="form-group">
+          <Field
+            type="text"
+            name="Description"
+            label="Description"
+            component={renderAreaField}
+          />
+        </div>
+        <div className="form-group">
+          <Field
+            type="text"
+            name="Email"
+            label="Email"
+            component={renderInputField}
+          />
+        </div>
       </div>
     )
   }
