+++ front/src/App/pages/SongRegister/index.js
@@ -4,16 +4,14 @@ import css from "./index.scss"
 import test1 from "src/assets/test_reg/jmb.png"
 import test2 from "src/assets/test_reg/pika.png"
 import TextInputForm from "./component/textInput"
-import ThumbNailInput from "./component/thumbnailInput"
+import RightSideForm from "./component/RightSideForm"
 const cx = classnames.bind(css)
 const moduleName = "SongRegist"
-//image selection은 input type hidden 의 value를 조정해서 submit하기.
 class SongRegist extends PureComponent {
   state = {
     selectedArtWork: `url(${test1})`
   }
   selectThumbnail = number => {
-    //나중에 reducer로 번호만 넘겨서 뒷단에서 thumbnail 매칭하기.
     switch (number) {
       case 1:
         this.setState(() => {
@@ -40,41 +38,27 @@ class SongRegist extends PureComponent {
         <div className={cx(`${moduleName}_top`)}>
           <h4>Register Song</h4>
         </div>
-        <div className={cx(`${moduleName}_body`)}>
-          <div className={cx(`${moduleName}_left`)}>
-            <TextInputForm />
-          </div>
-          <div className={cx(`${moduleName}_right`)}>
-            <div className={cx(`${moduleName}_right_top`)}>Album cover </div>
-            <div
-              className={cx(`${moduleName}_right_mid`)}
-              style={{ backgroundImage: this.state.selectedArtWork }}
-            />
-            <input
-              type="hidden"
-              name="thumbnail"
-              value={this.state.selectedArtWork}
-            />
-            <ThumbNailInput selectThumbnail={this.selectThumbnail} />
-            <div className={cx(`${moduleName}_right_bot_check`)}>
-              <label htmlFor="user_terms" className="checkbox-inline">
-                <input
-                  id="user_terms"
-                  type="checkbox"
-                  className="checkbox-inline"
-                />
-                I agree use this album cover.
-              </label>
+        <form>
+          <div className={cx(`${moduleName}_body`)}>
+            <div className={cx(`${moduleName}_left`)}>
+              <TextInputForm thumbnail={this.state.selectedArtWork} />
             </div>
-            <div className={cx(`${moduleName}_right_bot_submit`)}>
-              <button className="btn">Cancel</button>
-              <button className="btn">Register</button>
+            <div className={cx(`${moduleName}_right`)}>
+              <div className={cx(`${moduleName}_right_top`)}>Album cover </div>
+              <div
+                className={cx(`${moduleName}_right_mid`)}
+                style={{ backgroundImage: this.state.selectedArtWork }}
+              />
+              <RightSideForm selectThumbnail={this.selectThumbnail} />
+              <div className={cx(`${moduleName}_right_bot_submit`)}>
+                <button className="btn">Cancel</button>
+                <button className="btn">Register</button>
+              </div>
             </div>
           </div>
-        </div>
+        </form>
       </div>
     )
   }
 }
 export default SongRegist
-//export default SongRegist
